feat: Complete ESP32-S3-CAM system with AprilTag recognition and advanced features
Major Features Added:
• AprilTag Manager: Full AprilTag detection and tracking system
• Camera Mode Manager: Advanced camera operation modes and controls
• GPS Manager: Comprehensive GPS positioning and NMEA parsing
• EXIF GPS: GPS data embedding in image metadata
• MAVLink Manager: Complete MAVLink protocol support and message handling
• PSRAM Manager: Memory optimization for large image processing

Core System Enhancements:
• Enhanced camera management with power optimization
• Improved configuration system with validation
• Advanced upload management with retry logic
• Comprehensive power management across all modules
• Updated main sketch with new subsystem integration

Technical Improvements:
• Memory-efficient image processing for AprilTag detection
• Real-time GPS position tracking with accuracy validation
• MAVLink hardcoded protocol for embedded systems
• PSRAM utilization for high-resolution image capture
• Multi-mode camera operation (photo, video, recognition)

Hardware Support:
• ESP32-S3-CAM optimized pin configurations
• XIAO ESP32S3 compatibility
• Standard ESP32CAM support
• Multi-board GPIO abstraction

This implementation provides a complete computer vision and positioning system
for ESP32-S3-CAM applications including surveying, mapping, and autonomous systems.

+++ ESPCAMTRIP/upload_manager.cpp
@@ -127,7 +127,7 @@ void UploadManager::uploadPendingDirectories() {
           // Exponential backoff: 2^retry * 1000ms (1s, 2s, 4s, 8s...)
           uint32_t backoffDelay = (1 << retry) * 1000;
           // Cap at 30 seconds
-          backoffDelay = min(backoffDelay, 30000U);
+          backoffDelay = min(backoffDelay, (uint32_t)30000);
           Serial.printf("Waiting %u ms before retry...\n", backoffDelay);
           delay(backoffDelay);
         }
